+++ src/config/seatgroups.permission.php
@@ -10,7 +10,6 @@
 return [
     'seatgroups' => [
         'view',
-        'create',
-        'edit'
+        'create'
     ]
 ];