+++ src/resources/views/affiliation/partials/corporation-title-form.blade.php
@@ -7,6 +7,11 @@
     <label for="corporations">{{ trans('web::seat.available_corporations') }}</label>
     <select name="seatgroup-corporation-id" id="seatgroup-corporation-id" style="width: 100%" >
       <option></option>
+      @if(!$seatgroup->all_corporations)
+        @foreach($all_corporations_for_title as $corporation)
+          <option value="{{ $corporation->corporation_id }}">{{ $corporation->name }}</option>
+        @endforeach
+      @endif
     </select>
   </div>
 
@@ -20,7 +25,7 @@
   <div class="row">
     <div class="col-md-6"></div>
     <div class="form-group col-md-12">
-      <button type="submit" class="btn btn-success btn-block">Add Corporation</button>
+      <button type="submit" class="btn btn-success btn-block">Add Corporation Title Affiliation</button>
     </div>
   </div>
 </form>
@@ -51,28 +56,6 @@
       });
     }
 
-    function getCorporations() {
-      $('#seatgroup-corporation-id').empty();
-      $.ajax({
-        type   : 'POST',
-        url    : '{{ route('affiliation.get.corporation.list') }}',
-        data   : {
-          filter: 'corporation-tile-form',
-          seatgroup_id: '{{$seatgroup->id}}'
-        },
-        success: function (data) {
-          for (var i = 0; i < data.length; i++) {
-            $('#seatgroup-corporation-id').append($('<option></option>').attr('value', data[i].corporation_id).text(data[i].name));
-          }
-        },
-        error  : function (xhr, textStatus, errorThrown) {
-          console.log(xhr);
-          console.log(textStatus);
-          console.log(errorThrown);
-        }
-      })
-    }
-
     $("#seatgroup-corporation-id,"+ "#seatgroup-title-id").select2({
       placeholder: "{{ trans('web::seat.select_item_add') }}",
       allowClear: true
@@ -82,11 +65,6 @@
     $('#seatgroup-corporation-id').change(function(){
       getCorporationTitle();
     });
-
-    //TODO optimize this loading, so no corporation is preselected. This causes errors in network tab.
-    $(document).ready(function() {
-        getCorporations();
-    })
 
   </script>
 @endpush