+++ src/resources/views/partials/edit-button.blade.php
@@ -1,0 +1,5 @@
+<div class="btn-group pull-right">
+    <a href="{{ route('seatgroups.edit', $seatgroup->id)}}" class="btn btn-sm btn-warning pull-right">
+      <i class="fa fa-edit"></i> {{ trans('web::seat.edit') }}
+    </a>
+</div>